feature(Attributes): Renamed module
